feat: user_page 내 edit_profile 및 settings_page 생성 (#3)
user_page 내부에 edit_profile 페이지와 settings_page를 생성함. 각 페이지의 기본적인 구조를 구현하여 상세 페이지를 구성함.
