+++ fe/lib/pages/user/edit_profile_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class EditProfilePage extends StatelessWidget {
+  const EditProfilePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('정보 수정'),
+        backgroundColor: Colors.white,
+        foregroundColor: Colors.black,
+        elevation: 0,
+      ),
+      body: const Center(
+        child: Text('정보 수정 페이지'),
+      ),
+    );
+  }
+}
